Move unit tests to existing file; remove benchmark
