+++ test/test_cost.py
@@ -4,6 +4,8 @@
 # LICENSE file in the root directory of this source tree.
 
 import argparse
+import functools
+import operator
 import warnings
 from copy import deepcopy
 
@@ -97,6 +99,7 @@ from torchrl.objectives.utils import (
 )
 from torchrl.objectives.value.advantages import GAE, TD1Estimator, TDLambdaEstimator
 from torchrl.objectives.value.functional import (
+    _get_num_per_traj_init,
     generalized_advantage_estimate,
     td0_advantage_estimate,
     td1_advantage_estimate,
@@ -105,7 +108,13 @@ from torchrl.objectives.value.functional import (
     vec_td1_advantage_estimate,
     vec_td_lambda_advantage_estimate,
 )
-from torchrl.objectives.value.utils import _custom_conv1d, _make_gammas_tensor
+from torchrl.objectives.value.utils import (
+    _custom_conv1d,
+    _get_num_per_traj,
+    _inv_pad_sequence,
+    _make_gammas_tensor,
+    _split_and_pad_sequence,
+)
 
 
 class _check_td_steady:
@@ -4096,6 +4105,7 @@ class TestValues:
         r2 = generalized_advantage_estimate(
             gamma, lmbda, state_value, next_state_value, reward, done
         )
+
         torch.testing.assert_close(r1, r2, rtol=1e-4, atol=1e-4)
 
     @pytest.mark.parametrize("device", get_available_devices())
@@ -5133,6 +5143,117 @@ class TestBase:
 
         for key in ["module.1.bias", "module.1.weight"]:
             loss_module.module_b_params.flatten_keys()[key].requires_grad
+
+
+class TestUtils:
+    @pytest.mark.parametrize("B", [None, (1, ), (4, ), (2, 2, ), (1, 2, 8, )])  # fmt: skip
+    @pytest.mark.parametrize("T", [1, 10])
+    @pytest.mark.parametrize("device", get_available_devices())
+    def test_get_num_per_traj_no_stops(self, B, T, device):
+        """check _get_num_per_traj when input contains no stops"""
+        size = (*B, T) if B else (T,)
+
+        done = torch.zeros(*size, dtype=torch.bool, device=device)
+        splits = _get_num_per_traj(done)
+
+        count = functools.reduce(operator.mul, B, 1) if B else 1
+        res = torch.full((count,), T, device=device)
+
+        torch.testing.assert_close(splits, res)
+
+    @pytest.mark.parametrize("B", [(1, ), (3, ), (2, 2, ), (1, 2, 8, )])  # fmt: skip
+    @pytest.mark.parametrize("T", [5, 100])
+    @pytest.mark.parametrize("device", get_available_devices())
+    def test_get_num_per_traj(self, B, T, device):
+        """check _get_num_per_traj where input contains a stop at half of each trace"""
+        size = (*B, T)
+
+        done = torch.zeros(*size, dtype=torch.bool, device=device)
+        done[..., T // 2] = True
+        splits = _get_num_per_traj(done)
+
+        count = functools.reduce(operator.mul, B, 1)
+        res = [T - (T + 1) // 2 + 1, (T + 1) // 2 - 1] * count
+        res = torch.as_tensor(res, device=device)
+
+        torch.testing.assert_close(splits, res)
+
+    @pytest.mark.parametrize("B", [(1, ), (3, ), (2, 2, ), (1, 2, 8, )])  # fmt: skip
+    @pytest.mark.parametrize("T", [5, 100])
+    @pytest.mark.parametrize("device", get_available_devices())
+    def test_split_pad_reverse(self, B, T, device):
+        """calls _split_and_pad_sequence and reverts it"""
+        torch.manual_seed(42)
+
+        size = (*B, T)
+        traj = torch.rand(*size, device=device)
+        done = torch.zeros(*size, dtype=torch.bool, device=device).bernoulli(0.2)
+        splits = _get_num_per_traj(done)
+
+        splitted = _split_and_pad_sequence(traj, splits)
+        reversed = _inv_pad_sequence(splitted, splits).reshape(traj.shape)
+
+        torch.testing.assert_close(traj, reversed)
+
+    @pytest.mark.parametrize("B", [(1, ), (3, ), (2, 2, ), (1, 2, 8, )])  # fmt: skip
+    @pytest.mark.parametrize("T", [5, 100])
+    @pytest.mark.parametrize("device", get_available_devices())
+    def test_split_pad_no_stops(self, B, T, device):
+        """_split_and_pad_sequence on trajectories without stops should not change input but flatten it along batch dimension"""
+        size = (*B, T)
+        count = functools.reduce(operator.mul, size, 1)
+
+        traj = torch.arange(0, count, device=device).reshape(size)
+        done = torch.zeros(*size, dtype=torch.bool, device=device)
+
+        splits = _get_num_per_traj(done)
+        splitted = _split_and_pad_sequence(traj, splits)
+
+        traj_flat = traj.flatten(0, -2)
+        torch.testing.assert_close(traj_flat, splitted)
+
+    @pytest.mark.parametrize("device", get_available_devices())
+    def test_split_pad_manual(self, device):
+        """handcrafted example to test _split_and_pad_seqeunce"""
+
+        traj = torch.as_tensor([[0, 1, 2, 3, 4], [5, 6, 7, 8, 9]], device=device)
+        splits = torch.as_tensor([3, 2, 1, 4], device=device)
+        res = torch.as_tensor(
+            [[0, 1, 2, 0], [3, 4, 0, 0], [5, 0, 0, 0], [6, 7, 8, 9]], device=device
+        )
+
+        splitted = _split_and_pad_sequence(traj, splits)
+        torch.testing.assert_close(res, splitted)
+
+    @pytest.mark.parametrize("B", [(1, ), (3, ), (2, 2, ), (1, 2, 8, )])  # fmt: skip
+    @pytest.mark.parametrize("T", [5, 100])
+    @pytest.mark.parametrize("device", get_available_devices())
+    def test_split_pad_reverse_tensordict(self, B, T, device):
+        """calls _split_and_pad_sequence and reverts it on tensordict input"""
+        torch.manual_seed(42)
+
+        td = TensorDict(
+            {
+                "observation": torch.arange(T, dtype=torch.float32, device=device)
+                .unsqueeze(-1)
+                .expand(*B, T, 3),
+                "is_init": torch.zeros(
+                    *B, T, 1, dtype=torch.bool, device=device
+                ).bernoulli(0.3),
+            },
+            [*B, T],
+            device=device,
+        )
+
+        is_init = td.get("is_init").squeeze(-1)
+        splits = _get_num_per_traj_init(is_init)
+        splitted = _split_and_pad_sequence(
+            td.select("observation", strict=False), splits
+        )
+
+        reversed = _inv_pad_sequence(splitted, splits)
+        reversed = reversed.reshape(td.shape)
+        torch.testing.assert_close(td["observation"], reversed["observation"])
 
 
 @pytest.mark.parametrize("updater", [HardUpdate, SoftUpdate])
